refactor(telegram): extract inbound ingest/event building into TelegramIngestMixin (adapter god-file slice)
